delete sync dao
