+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/UserSessionScope.kt
@@ -575,7 +575,6 @@ class UserSessionScope internal constructor(
 
     private val userRepository: UserRepository = UserDataSource(
         userStorage.database.userDAO,
-        userStorage.database.syncDAO,
         userStorage.database.metadataDAO,
         userStorage.database.clientDAO,
         authenticatedNetworkContainer.selfApi,
