+++ persistence/src/commonMain/db_user/com/wire/kalium/persistence/Users.sq
@@ -160,3 +160,6 @@ UPDATE User SET name = ? WHERE qualified_id = ?;
 selectUsersWithoutMetadata:
 SELECT * FROM User AS user
 WHERE deleted = 0 AND incomplete_metadata = 1;
+
+userIdsWithoutSelf:
+SELECT qualified_id FROM User WHERE qualified_id != (SELECT id FROM SelfUser);
