+++ persistence/src/commonMain/kotlin/com/wire/kalium/persistence/db/UserDatabaseBuilder.kt
@@ -36,8 +36,6 @@ import com.wire.kalium.persistence.dao.PrekeyDAO
 import com.wire.kalium.persistence.dao.PrekeyDAOImpl
 import com.wire.kalium.persistence.dao.ServiceDAO
 import com.wire.kalium.persistence.dao.ServiceDAOImpl
-import com.wire.kalium.persistence.dao.SyncDAO
-import com.wire.kalium.persistence.dao.SyncDAOImpl
 import com.wire.kalium.persistence.dao.TeamDAO
 import com.wire.kalium.persistence.dao.TeamDAOImpl
 import com.wire.kalium.persistence.dao.UserDAO
@@ -194,8 +192,6 @@ class UserDatabaseBuilder internal constructor(
     val metadataDAO: MetadataDAO
         get() = MetadataDAOImpl(database.metadataQueries, metadataCache, databaseScope, queriesContext)
 
-    val syncDAO: SyncDAO
-        get() = SyncDAOImpl(database.syncQueries, queriesContext)
     val clientDAO: ClientDAO
         get() = ClientDAOImpl(database.clientsQueries, queriesContext)
